+++ src/thinger/utils/date.h
@@ -5,18 +5,54 @@
 #include <sstream> // stringstream
 #include <iomanip> // put_time
 
-namespace Date {
+namespace Date { // TODO: please, think. This file is kind of a mess. Treat it as objects that can undergo format changes maintaining the same reference, and not different objects and static methods each time
 
-    static std::string now_iso8601(const bool extended = false) {
+    static std::string now_iso8601(const char del = '-', const bool extended = false) {
         auto now = std::chrono::system_clock::now();
         auto in_time_t = std::chrono::system_clock::to_time_t(now);
 
+        //std::string format = "%Y"%m%d";
+        //std::string format = std::string("%Y").append(&del);//"%m%d";
+        std::string format =
+            std::string("%Y").append(&del)+
+            std::string("%m").append(&del)+
+            std::string("%d");
+
         std::stringstream ss;
-        if (!extended)
-            ss << std::put_time(std::localtime(&in_time_t), "%Y-%m-%d");
-        // TODO
+        if (extended)
+            format =
+                std::string("%Y").append(&del)+
+                std::string("%m").append(&del)+
+                "%dT%H%M%SZ";
+            //ss << std::put_time(std::localtime(&in_time_t), "%Y".append(del)+"%m"+append(del)+"%d");
+        ss << std::put_time(std::localtime(&in_time_t), format.c_str());
+        // TODO: iso8601 extended standard
         //else
-        //    ss << std::put_time(std::localtime(&in_time_t), "%Y-%m-%dT%X%z);
+
+        return ss.str();
+    }
+
+    static std::string now_utc_iso8601(const char del = '-', const bool extended = false) {
+        auto now = std::chrono::system_clock::now();
+        auto in_time_t = std::chrono::system_clock::to_time_t(now);
+
+        //std::string format = "%Y"%m%d";
+        //std::string format = std::string("%Y").append(&del);//"%m%d";
+        std::string format =
+            std::string("%Y").append(&del)+
+            std::string("%m").append(&del)+
+            std::string("%d");
+
+        std::stringstream ss;
+        if (extended)
+            format =
+                std::string("%Y").append(&del)+
+                std::string("%m").append(&del)+
+                "%dT%H%M%SZ";
+            //ss << std::put_time(std::localtime(&in_time_t), "%Y".append(del)+"%m"+append(del)+"%d");
+        ss << std::put_time(std::gmtime(&in_time_t), format.c_str());
+        // TODO: iso8601 extended standard
+        //else
 
         return ss.str();
     }